django/upload_json/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.14.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pkgutil-resolve-name==1.3.10
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
ua-parser==0.16.0
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
zipp==3.8.1
